+++ lib/include/zuc_internal.h
@@ -918,13 +918,13 @@ IMB_DLL_LOCAL void
 asm_ZucGenKeystream_16_avx512(ZucState16_t *pState,
                               uint32_t *pKstr,
                               const unsigned key_off,
-                              const uint64_t numRounds);
+                              const uint32_t numRounds);
 
 IMB_DLL_LOCAL void
 asm_ZucGenKeystream_16_gfni_avx512(ZucState16_t *pState,
                                    uint32_t *pKstr,
                                    const unsigned key_off,
-                                   const uint64_t numRounds);
+                                   const uint32_t numRounds);
 
 /**
  ******************************************************************************
@@ -959,14 +959,14 @@ asm_ZucGenKeystream_16_skip8_avx512(ZucState16_t *pState,
                                     uint32_t *pKstr,
                                     const unsigned key_off,
                                     const uint16_t lane_mask,
-                                    const uint64_t numRounds);
+                                    const uint32_t numRounds);
 
 IMB_DLL_LOCAL void
 asm_ZucGenKeystream_16_skip8_gfni_avx512(ZucState16_t *pState,
                                          uint32_t *pKstr,
                                          const unsigned key_off,
                                          const uint16_t lane_mask,
-                                         const uint64_t numRounds);
+                                         const uint32_t numRounds);
 /**
  ******************************************************************************
  *
@@ -1205,31 +1205,31 @@ IMB_DLL_LOCAL uint32_t asm_Eia3RemainderAVX(const void *ks, const void *data,
  *****************************************************************************/
 IMB_DLL_LOCAL void asm_Eia3RemainderAVX512(uint32_t *T, const void *ks,
                                            const void *data,
-                                           const uint64_t n_bits);
+                                           const uint32_t n_bits);
 
 IMB_DLL_LOCAL uint32_t asm_Eia3RemainderAVX512_16(uint32_t *T,
                                                   const uint32_t *ks,
                                                   const void **data,
                                                   uint16_t *lens,
-                                                  const uint64_t commonBits);
+                                                  const uint32_t commonBits);
 
 IMB_DLL_LOCAL uint32_t asm_Eia3_256_RemainderAVX512_16(uint32_t *T,
                                                   const uint32_t *ks,
                                                   const void **data,
                                                   uint16_t *lens,
-                                                  const uint64_t commonBits);
+                                                  const uint32_t commonBits);
 
 IMB_DLL_LOCAL uint32_t asm_Eia3RemainderAVX512_16_VPCLMUL(uint32_t *T,
                                                   const uint32_t *ks,
                                                   const void **data,
                                                   uint16_t *lens,
-                                                  const uint64_t commonBits);
+                                                  const uint32_t commonBits);
 
 IMB_DLL_LOCAL uint32_t asm_Eia3_256_RemainderAVX512_16_VPCLMUL(uint32_t *T,
                                                   const uint32_t *ks,
                                                   const void **data,
                                                   uint16_t *lens,
-                                                  const uint64_t commonBits);
+                                                  const uint32_t commonBits);
 
 IMB_DLL_LOCAL
 void zuc_eia3_4_buffer_job_gfni_sse(const void * const pKey[4],
